Replace murmur2hash for getting stack ids
Instead of utilizing our own hash function
for generating unique stack ids (as this inherently
leads to hash collisions due to the birthday
paradox) use a BPF map to store stacks
as keys (with the id being the value).

The tradeoff here is that we're using more memory
(we need another large map to track the stack ids)
and it's potentially slower as it may result in
a map lookup and an update. However it's more
accurate, which is probably what bpftrace should
be optimizing for.

This wasn't done when murmur2hash was originally
added because we supported a kernel version that
had limits on the size of the stack key; this
is no longer the case.

Signed-off-by: Jordan Rome <linux@jordanrome.com>

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/4860, branch: user/jordalgo/cond_stdlib_import/3
Signed-off-by: Jordan Rome <linux@jordanrome.com>